pass SE3 tests
