+++ geomstats/geometry/matrices.py
@@ -96,7 +96,6 @@ class Matrices(Euclidean):
         """
         is_vectorized = (gs.ndim(gs.array(mat)) == 3)
         axes = (0, 2, 1) if is_vectorized else (1, 0)
-        print(mat, is_vectorized, axes)
         return gs.transpose(mat, axes)
 
     @classmethod
